Update first-time-travel-complete-concur-profile.md
I believe that the existing description is outdated (the screenshots provided in "Having Trouble? / Check out the screenshots" show an older version of Concur where "I want to..." no longer exists). I have updated step 2 to match the current experience (and also added a period to step 3 for consistency with step 1, 4, and 5.